move more components closer to the place of usage
