+++ extension/src/panel/routes/Edit/wrapRequest.ts
@@ -1,7 +1,6 @@
+import { LegacyConnection, TransactionData } from '@/types'
 import { ContractFactories, KnownContracts } from '@gnosis.pm/zodiac'
-
 import { MetaTransactionData } from '@safe-global/safe-core-sdk-types'
-import { LegacyConnection, TransactionData } from '../../types'
 
 const RolesV1Interface =
   ContractFactories[KnownContracts.ROLES_V1].createInterface()
